refactor: split apiHandler.test.js into focused test files and document Playwright E2E patterns
Split the 1798-line tests/api/apiHandler.test.js into seven smaller files
under tests/api/apiHandler/, organised by test concern:

- shared.js: All shared constants and helper functions (imported by all test files)
- registry.test.js: Allowlisted method handler registry tests
- globalExposure.test.js: Trailing-underscore global exposure tests in VM context
- dispatcher-core.test.js: Core dispatch, request validation, request ID generation
- dispatcher-routing.test.js: Method routing (reference data, transport helpers, upsert)
- dispatcher-envelope-errors.test.js: IN_USE and INVALID_REQUEST envelope error cases
- dispatcher-errors.test.js: Error handling, boundary logging, error type mapping
- dispatcher-integration.test.js: Real handler integration, VM context, delegation

Each file is self-contained with its own describe/beforeEach/afterEach and
under 500 lines, resolving the pre-existing max-lines lint regression.

Also documented E2E Runtime Mock Infrastructure patterns in
frontend-testing.md covering RuntimeScenario, response queues, StrictMode-safe
sizing, deferred responses, and antd Select interaction.
